89389- Shift messageBuilder to Common project AB#89389 AB#90996
