+++ src/UKHO.ERPFacade.Common/Services/IRecordOfSaleSapMessageBuilder.cs
@@ -1,7 +1,7 @@
 ﻿using System.Xml;
 using UKHO.ERPFacade.Common.Models;
 
-namespace UKHO.ERPFacade.API.Helpers
+namespace UKHO.ERPFacade.Common.Services
 {
     public interface IRecordOfSaleSapMessageBuilder
     {
