+++ tests/UKHO.ERPFacade.Common.UnitTests/UKHO.ERPFacade.Common.UnitTests.csproj
@@ -16,7 +16,7 @@
   <ItemGroup>
     <PackageReference Include="FakeItEasy" Version="7.3.1" />
     <PackageReference Include="FluentAssertions" Version="6.10.0" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.1.0" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.2" />
     <PackageReference Include="NUnit" Version="3.13.3" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
     <PackageReference Include="NUnit.Analyzers" Version="3.3.0" />
@@ -28,6 +28,13 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\UKHO.ERPFacade.Common\UKHO.ERPFacade.Common.csproj" />
+    <ProjectReference Include="..\UKHO.ERPFacade.API.UnitTests\UKHO.ERPFacade.API.UnitTests.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="ERPTestData\NewLicencePayloadTest.xml">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
 </Project>
